pyproject.toml
[tool.poetry]
name = "cowtrack"
version = "0.1.0"
description = ""
authors = ["cowtrack"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
django-debug-toolbar = "^4.2.0"
django-money = "^3.3.0"
mysqlclient = "^2.2.0"
djangorestframework = "^3.14.0"
drf-nested-routers = "^0.93.4"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
ujson = "^5.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
